exchange: make the sketch conversion glue format-neutral
The drawing→sketch and sketch→drawing converters were typed on dwg.* and lived
in dwg-named files, even though they are pure geometry. Re-type them on the
neutral kernel/exchange/drawing model and move them into format-agnostic files so
the upcoming DXF importer/exporter reuses them instead of duplicating:

- sketch_from_drawing.go / sketch_from_drawing3d.go: drawing.Entity → 2D/3D
  sketch, plus the shared importDrawing orchestrator (scale to db units, choose
  2D-plane vs 3D sketch) and drawingToDocumentScale (renamed from
  dwgToDocumentScale).
- sketch_to_drawing.go: sketch → []drawing.Entity (renamed from sketchToDWG).

dwgimport.go / dwgexport.go shrink to the DWG-specific entry points (ImportDWG
wraps dwg.Decode + importDrawing; ExportDWG wraps sketchToDrawing + dwg.Write).
Public names (ImportDWG/ExportDWG/DWGImportResult) and behaviour are unchanged;
the exchange test suite passes.
